% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_to_geojson.R
\name{excel_to_geojson}
\alias{excel_to_geojson}
\title{This function reads an Excel file, cleans the column names,
converts a specified GeoJSON geometry column to an sf object, and returns it.}
\usage{
excel_to_geojson(data)
}
\arguments{
\item{data}{A string specifying the path to the Excel file.}
}
\value{
An sf object containing the transformed spatial data.
}
\description{
This function reads an Excel file, cleans the column names,
converts a specified GeoJSON geometry column to an sf object, and returns it.
}
